Move the TOML creation out of the project command.
